requirements.txt
langchain>=0.0.267
faiss-cpu>=1.7.4
chromadb>=0.4.13
openai>=1.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
numpy>=1.24.3
pandas>=2.0.3
pydantic>=2.4.2
pytest>=7.4.0
sentence-transformers>=2.2.2
torch>=2.0.1
transformers>=4.33.2
tqdm>=4.66.1
